[uishell] Add UIEvent.ShowSummary
