+++ uishell/src/main/scala/net/kemitix/throp/uishell/UIEvent.scala
@@ -1,9 +1,13 @@
 package net.kemitix.throp.uishell
 
+import net.kemitix.thorp.domain.Counters
+
 sealed trait UIEvent
 object UIEvent {
   case object ShowValidConfig extends UIEvent
 
   case class RemoteDataFetched(size: Int) extends UIEvent
+
+  case class ShowSummary(counters: Counters) extends UIEvent
 
 }
