+++ uishell/src/main/scala/net/kemitix/throp/uishell/UIShell.scala
@@ -3,6 +3,7 @@ package net.kemitix.throp.uishell
 import net.kemitix.eip.zio.MessageChannel
 import net.kemitix.thorp.config.Config
 import net.kemitix.thorp.console.{Console, ConsoleOut}
+import net.kemitix.thorp.domain.Terminal.eraseToEndOfScreen
 import zio.UIO
 
 object UIShell {
@@ -19,6 +20,12 @@ object UIShell {
           } yield ()
         case UIEvent.RemoteDataFetched(size) =>
           Console.putStrLn(s"Found $size remote objects")
+        case UIEvent.ShowSummary(counters) =>
+          Console.putStrLn(eraseToEndOfScreen) *>
+            Console.putStrLn(s"Uploaded ${counters.uploaded} files") *>
+            Console.putStrLn(s"Copied   ${counters.copied} files") *>
+            Console.putStrLn(s"Deleted  ${counters.deleted} files") *>
+            Console.putStrLn(s"Errors   ${counters.errors}")
       }
     }
 }
